Bugfix related to QT Calculation

+++ include/host/HostSideComputation.hpp
@@ -10,21 +10,21 @@ namespace HostSideComputation {
 
     void PrecomputeStatistics(std::array<double, n> &T, std::array<data_t, n - m + 1> &QT, std::array<ComputePack, n - m + 1> &data) {
         // Calculate the initial mean, then update using moving mean.
-        double mean = std::accumulate(T.begin(), T.begin() + m, 0); mean /= m;
+        double mean = std::accumulate(T.begin(), T.begin() + m, (double)0); mean /= m;
         double prev_mu, mu0 = mean;
 
         // Compute Statistics in single pass through the data
         for (index_t i = 0; i < n - m + 1; ++i) {
             prev_mu = mean;
-            mean += (i > 0) ? ((T[i + m - 1] - T[i - 1]) / m) 
-                            : 0;
+            mean += (i > 0) ? ((T[i + m - 1] - T[i - 1]) / m) : 0;
             data[i].df = (i > 0) ? static_cast<data_t>((T[i + m - 1] - T[i - 1]) / 2)
                             : static_cast<data_t>(0);
             data[i].dg = (i > 0) ? static_cast<data_t>(((T[i + m - 1] - mean) + (T[i - 1] - prev_mu))) 
                             : static_cast<data_t>(0);
-            double qt = 0, inv = 0;
+
+	    double qt = 0, inv = 0;
             for (index_t k = 0; k < m; ++k) {
-                qt  += (T[i + k] - mean) * (T[k] - mu0);
+                qt += (T[i + k] - mean) * (T[k] - mu0);
                 inv += (T[i + k] - mean) * (T[i + k] - mean);
             }
             QT[i] = qt;
